Clean-up `FullscreenManager` and add spec references https://bugs.webkit.org/show_bug.cgi?id=288285 rdar://145351421
Reviewed by Alex Christensen.

This PR does not contain any C++ code changes, and purely moves code around.

- Re-orders methods into sections that clearly correspond to spec algorithms:
   - Enter fullscreen flow is now grouped in one section
   - Exit fullscreen flow is now grouped in one section
   - Event dispatching methods are now grouped
- Adds spec references.
- Moves code less relevant to the understanding of the code to the header file.

This will make the code easier to read and cross-reference with the spec, especially after it is refactored for site isolation.

* Source/WebCore/dom/FullscreenManager.cpp:
(WebCore::FullscreenManager::isFullscreenEnabled const):
(WebCore::FullscreenManager::willEnterFullscreen):
(WebCore::FullscreenManager::didEnterFullscreen):
(WebCore::FullscreenManager::isSimpleFullscreenDocument const):
(WebCore::FullscreenManager::exitRemovedFullscreenElement):
(WebCore::FullscreenManager::cancelFullscreen):
(WebCore::FullscreenManager::queueFullscreenChangeEventForDocument):
(WebCore::FullscreenManager::emptyEventQueue):
(): Deleted.
* Source/WebCore/dom/FullscreenManager.h:

Canonical link: https://commits.webkit.org/290874@main
